requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
dnspython==2.4.2
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
MarkupSafe==2.1.3
neo4j==5.15.0
numpy==1.26.2
psycopg2-binary==2.9.9
pymongo==4.6.1
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
threadpoolctl==3.2.0
urllib3==2.1.0
Werkzeug==3.0.1
